added appDeliveryData to getCompleteDownloadInfo
Hi,
first off, great job on version 4! 
customising the useragent is a great addition and my team will defiantly use it.

this PR will return appDeliveryData with the url and cookie jar so we can use it to further enrich our apk data.
i rather not call `Api.deliveryData` since it will trigger another API call - and we need to reduce the number of calls.

if you think we should implement a memoization for that request instead of adding that info to this function let me know.
Thanks!